refactor: ParsedAnnouncement를 application/model 패키지로 이동
