+++ eeos/src/main/java/com/blackcompany/eeos/announcement/application/support/AnnouncementParser.java
@@ -1,4 +1,5 @@
 package com.blackcompany.eeos.announcement.application.support;
+import com.blackcompany.eeos.announcement.application.model.ParsedAnnouncement;
 
 public interface AnnouncementParser {
 	ParsedAnnouncement parse(String text);
