+++ eeos/src/main/java/com/blackcompany/eeos/announcement/application/support/AnnouncementTextParser.java
@@ -1,4 +1,5 @@
 package com.blackcompany.eeos.announcement.application.support;
+import com.blackcompany.eeos.announcement.application.model.ParsedAnnouncement;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
